refactor: 패키지 구조, 클래스 네이밍 수정
